Add generic traversals.
This change has some odd effects. It measurably increases allocation per path
attempted, resulting in as much as 1/3 more allocation in the exhaustive
traversals.

But it does not measurably harm performance. Which is unexpected.

I can't explain the allocation behavior so far, but I like this change
architecturally, so I'll go ahead and risk the performance hit.